+++ src/util.jl
@@ -71,8 +71,8 @@ function fgen(name::Symbol, ar::Int, s1::SorX)
     fgen(name, ar, s1, s1, s1, s1)
 end
 
-function recur_active!(active::BitArray, ind::Int8, xs::Array{Int8},
-                       ys::Array{Int8}, fs::Array{Int8},
+function recur_active!(active::BitArray, ind::Int16, xs::Array{Int16},
+                       ys::Array{Int16}, fs::Array{Int16},
                        two_arity::BitArray)::Nothing
     if ind > 0 && ~active[ind]
         active[ind] = true
@@ -83,16 +83,16 @@ function recur_active!(active::BitArray, ind::Int8, xs::Array{Int8},
     end
 end
 
-function find_active(cfg::Dict, genes::Array{Int8},
-                     outputs::Array{Int8})::BitArray
+function find_active(cfg::Dict, genes::Array{Int16},
+                     outputs::Array{Int16})::BitArray
     R = cfg["rows"]
     C = cfg["columns"]
     active = falses(R, C)
-    xs = genes[:, :, 1] .- Int8(cfg["n_in"])
-    ys = genes[:, :, 2] .- Int8(cfg["n_in"])
+    xs = genes[:, :, 1] .- Int16(cfg["n_in"])
+    ys = genes[:, :, 2] .- Int16(cfg["n_in"])
     fs = genes[:, :, 3]
     for i in eachindex(outputs)
-        recur_active!(active, outputs[i] - Int8(cfg["n_in"]), xs, ys, fs,
+        recur_active!(active, outputs[i] - Int16(cfg["n_in"]), xs, ys, fs,
                       cfg["two_arity"])
     end
     active
